Fix life steal type mismatch - convert healAmount to Int
The healAmount was a Float (damage * percent), but playerHp and maxHp
are Ints. This caused kotlin.math.min() to fail with type ambiguity.

+++ NeonSurvivor/app/src/main/java/com/example/neonsurvivor/GameView.kt
@@ -1349,7 +1349,7 @@ class GameView(context: Context) : View(context) {
 
                     // Life Steal - heal based on damage dealt
                     if (lifeStealPercent > 0f) {
-                        val healAmount = finalDamage * lifeStealPercent
+                        val healAmount = (finalDamage * lifeStealPercent).toInt()
                         playerHp = kotlin.math.min(playerHp + healAmount, maxHp)
                     }
